Move layout components to the right folder
