+++ src/components/Layout/components/Header/components/ResponsiveLogo/ResponsiveLogo.tsx
