+++ src/components/Layout/components/Header/components/ResponsiveLogo/index.ts
