+++ src/components/Layout/components/Header/components/ResponsiveLogo/styles.tsx
