+++ src/components/Layout/components/MainContainer/styles.tsx
